chore: remove previous implementation approach
